Create batch_tcpdump
This batch file runs a tcpdump off the firewall by running an AutoHotkey script. Then when that is finished, it runs Wireshark and tshark in Wireshark to take the .pcap to .csv. It emails to let you know the batch has started, tcpdump has finished and the file exists and the conversion to .csv exists also. (Must have Blat installed - http://www.blat.net/)